Renamed LICENSE.md to LICENSE.txt; added link.
The Apache 2.0 license is in text format and renders somewhat funny when
interpreted as Markdown. This makes it clear that it's a plain text file.

Also added a reference to the license file from README so that new users can
easily tell what license this repo is under and easily click through to see the
full license text.